requirements.txt
pillow
qrcode
requests
pycrypto